backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.0
python-dotenv==1.0.0
flask-jwt-extended==4.6.0